# Speed up model prediction tests by not setting up both nets every time