Before 48.94 seconds 
```
(venv) nthomas (master) castle $ PYTHONPATH=. py.test
============================================================================================================ test session starts ============================================================================================================
platform darwin -- Python 3.6.2, pytest-3.2.3, py-1.4.34, pluggy-0.4.0
rootdir: /Users/nthomas/git/castle, inifile:
plugins: cov-2.5.1
collected 11 items

tests/test_net.py ...
tests/test_tree.py ........

======================================================================================================== 11 passed in 48.94 seconds =========================================================================================================
```

After 14.45 seconds
```
(venv) nthomas (master *) castle $ PYTHONPATH=. py.test
============================================================================================================ test session starts ============================================================================================================
platform darwin -- Python 3.6.2, pytest-3.2.3, py-1.4.34, pluggy-0.4.0
rootdir: /Users/nthomas/git/castle, inifile:
plugins: cov-2.5.1
collected 11 items

tests/test_net.py ...
tests/test_tree.py ........

======================================================================================================== 11 passed in 14.45 seconds =========================================================================================================

```